% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ga_service_info}
\alias{ga_service_info}
\title{GA Service Information}
\format{A data frame with 1 observations and 14 variables \describe{
\item{Name}{Product Name}
\item{Short_Name}{Product short name}
\item{Service}{Product code}
\item{xmin}{left bounding longitude in decimal degrees}
\item{xmax}{right bounding longitude in decimal degrees}
\item{ymin}{bottom latitude in decimal degrees}
\item{ymax}{top bounding latitude in decimal degrees}
\item{offset_x}{Cell resolution in x dimension}
\item{offset_y}{Cell resolution in y dimension}
\item{origin_x}{x coordinate result of
  \code{\link[raster:origin]{raster::origin()}} for this dataset.}
\item{origin_y}{y coordinate result of
  \code{\link[raster:origin]{raster::origin()}} for this dataset.}
\item{ncol}{number of raster cells in x dimension}
\item{nrow}{number of raster cells in y dimension}
\item{na_val}{Raster No Data value}
}}
\usage{
ga_service_info
}
\description{
A data frame containing information about WCS services available from
GeoScience Australia.
}
\keyword{datasets}
