% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosciausws.R
\docType{package}
\name{geosciausws-package}
\alias{geosciausws}
\alias{geosciausws-package}
\title{geosciausws: Geoscience Australia services}
\description{
Provides access to Geoscience Australia raster datasets via existing
WCS services.
}
\author{
\strong{Maintainer}: Lauren O'Brien \email{obrlsoilau@gmail.com}

}
