% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url_generate.R
\name{make_ga_url}
\alias{make_ga_url}
\title{Make GeoScience Australia URL}
\usage{
make_ga_url(product = NULL, aoi = NULL, req_type = "cov")
}
\arguments{
\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[geosciausws:ga_service_info]{ga_service_info}}.}

\item{aoi}{Vector of WGS84 coordinates defining a rectangular area of
interest. The vector may be specified directly in the order xmin, ymin,
xmax, ymax, or the function can derive an aoi from the boundary of an `sf`
or `raster` object.}

\item{req_type}{Character; one of 'cap', 'cov' or 'desc'. Defaults to 'cov'.}
}
\description{
Generate the WCS GetCoverage URL for data products available from
Geoscience Australia.
}
\keyword{internal}
