% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{metadata_ga}
\alias{metadata_ga}
\title{download GA web service metadata}
\usage{
metadata_ga(product = NULL, req_type = "desc", format = "native")
}
\arguments{
\item{product}{Character, one of the options from column 'Short_Name' in
\code{\link[geosciausws:ga_service_info]{ga_service_info}}.}

\item{req_type}{Character; one of 'cap' or 'desc'. Defaults to 'desc'.}

\item{format}{Character; one of 'xml' or 'native'. Defaults to 'native'.}
}
\value{
A list or xml document object, depending on the value of 'format'.
}
\description{
Retrieves metadata from Geoscience Australia WCS
endpoints in XML or list format.
}
\note{
WCS services < v 2.0 can only return XML formatted data; JSON is not an
  option.
}
